+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added development field support
 - Added Branch creation and Rovo Dev actions in Create Issue Page
+- Added filter by project support
 
 ## What's new in 4.0.9